+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 This project uses [Semantic Versioning 2.0.0](http://semver.org/).
 
-#### master
+#### 4.1.0
 
 - NEW: Added domain premium price support (dnsimple/dnsimple-ruby#143)